02_LLM/socket/requirements.txt
bcrypt==4.0.1
cryptography
fastapi
uvicorn[standard]
psycopg2-binary
PyMySQL
python-jose
python-multipart
SQLAlchemy
uvicorn
passlib==1.7.4
pytest
httpx
pytest-asyncio
aiofiles
jinja2
uvicorn
Jinja2
mangum
openai
python-multipart
psutil
websocket-client==1.7.0
locust==2.24.0
aiohttp==3.9.3
asyncio==3.4.3
aioredis==2.0.1
aiokafka==0.8.1
redis==5.0.1
kafka-python==2.0.2
prometheus-client==0.19.0 
alembic